refactor(ui): extract ResourceDetailContent for lazy loading
